Moved _build_custom_neural_network
